added method PeerManager.setName to alow to rename a peer #317
