+++ management/server/subutai-common/src/main/java/io/subutai/common/peer/PeerInfo.java
@@ -10,7 +10,6 @@ import org.codehaus.jackson.annotate.JsonIgnore;
 import org.codehaus.jackson.annotate.JsonProperty;
 
 import io.subutai.common.settings.Common;
-import io.subutai.common.settings.SystemSettings;
 
 
 /**
@@ -76,7 +75,7 @@ public class PeerInfo implements Serializable
 
     public String getName()
     {
-        return String.format( "Peer %s on %s ", id, ip );
+        return name;
     }
